Conform to PSR-2, add Composer/PHPUnit configs
Update Makefile, add SETUP.md
Add PHP CodeSniffer
Use @group annotation to skip tests
