+++ .gitignore
@@ -2,3 +2,5 @@
 .DS_Store
 tmp
 *.phar
+vendor
+composer.lock
