+++ Clock/ClockTest.php
@@ -1,8 +1,7 @@
 <?php
+namespace Exercism\Clock;
 
-require_once realpath(__DIR__ . '/clock.php');
-
-class ClockTest extends PHPUnit_Framework_TestCase
+class ClockTest extends \PHPUnit_Framework_TestCase
 {
     public function testOnTheHour()
     {
@@ -11,19 +10,17 @@ class ClockTest extends PHPUnit_Framework_TestCase
         $this->assertEquals('08:00', $clock->__toString());
     }
 
+    /** @group skipped */
     public function testPastTheHour()
     {
-        $this->markTestSkipped();
-
         $clock = new Clock(11, 9);
 
         $this->assertEquals('11:09', $clock->__toString());
     }
 
+    /** @group skipped */
     public function testAddingAFewMinutes()
     {
-        $this->markTestSkipped();
-
         $clock = new Clock(10);
 
         $clock = $clock->add(3);
@@ -31,10 +28,9 @@ class ClockTest extends PHPUnit_Framework_TestCase
         $this->assertEquals('10:03', $clock->__toString());
     }
 
+    /** @group skipped */
     public function testAddingOverAnHour()
     {
-        $this->markTestSkipped();
-
         $clock = new Clock(10);
 
         $clock = $clock->add(61);
@@ -42,10 +38,9 @@ class ClockTest extends PHPUnit_Framework_TestCase
         $this->assertEquals('11:01', $clock->__toString());
     }
 
+    /** @group skipped */
     public function testWrapAroundAtMidnight()
     {
-        $this->markTestSkipped();
-
         $clock = new Clock(23, 30);
 
         $clock = $clock->add(60);
@@ -53,10 +48,9 @@ class ClockTest extends PHPUnit_Framework_TestCase
         $this->assertEquals('00:30', $clock->__toString());
     }
 
+    /** @group skipped */
     public function testSubtractMinutes()
     {
-        $this->markTestSkipped();
-
         $clock = new Clock(10);
 
         $clock = $clock->sub(90);
@@ -64,10 +58,9 @@ class ClockTest extends PHPUnit_Framework_TestCase
         $this->assertEquals('08:30', $clock->__toString());
     }
 
+    /** @group skipped */
     public function testWrapAroundBackwards()
     {
-        $this->markTestSkipped();
-
         $clock = new Clock(0, 30);
 
         $clock = $clock->sub(60);
@@ -75,10 +68,9 @@ class ClockTest extends PHPUnit_Framework_TestCase
         $this->assertEquals('23:30', $clock->__toString());
     }
 
+    /** @group skipped */
     public function testWrapAroundDay()
     {
-        $this->markTestSkipped();
-
         $clock = new Clock(5, 32);
 
         $clock = $clock->add(25 * 60);
@@ -86,10 +78,9 @@ class ClockTest extends PHPUnit_Framework_TestCase
         $this->assertEquals('06:32', $clock->__toString());
     }
 
+    /** @group skipped */
     public function testWrapAroundDayBackwards()
     {
-        $this->markTestSkipped();
-
         $clock = new Clock(5, 32);
 
         $clock = $clock->sub(25 * 60);
@@ -97,17 +88,15 @@ class ClockTest extends PHPUnit_Framework_TestCase
         $this->assertEquals('04:32', $clock->__toString());
     }
 
+    /** @group skipped */
     public function testEquivalentClocks()
     {
-        $this->markTestSkipped();
-
         $this->assertEquals(new Clock(15, 37), new Clock(15, 37));
     }
 
+    /** @group skipped */
     public function testInequivalentClocks()
     {
-        $this->markTestSkipped();
-
         $this->assertNotEquals(new Clock(01, 01), new Clock(18, 32));
     }
 }
