+++ Clock/Example.php
@@ -1,4 +1,5 @@
 <?php
+namespace Exercism\Clock;
 
 class Clock
 {
