+++ Wordy/ArgumentError.php
@@ -1,0 +1,6 @@
+<?php
+namespace Exercism\Wordy;
+
+class ArgumentError extends \InvalidArgumentException
+{
+}
