+++ composer.json
@@ -1,0 +1,10 @@
+{
+    "require-dev": {
+        "phpunit/phpunit": "4.1.*"
+    },
+    "autoload": {
+        "psr-4": {
+            "Exercism\\": ""
+        }
+    }
+}
